+++ CHANGES.md
@@ -5,6 +5,7 @@
 # Next
 
 - Fix behaviour of hot-swaping with dynamic composition (replace defs)
+- Include action input helper by default
 
 WIP: